emp/serializers.py
from rest_framework import serializers
from .models import *


class DepartmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Department
        fields = '__all__'

class EmployeeSerializer(serializers.ModelSerializer):
    dept=DepartmentSerializer(many=True,read_only=True)
    class Meta:
        model = Employee
        fields = ["id",'name','salary','department','dept']
        

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = '__all__'
